requirements.txt
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.11
idna==2.10
numpy==1.19.5
pandas==1.1.5
psycopg2==2.8.6
PyGithub==1.54.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
urllib3==1.26.3
wrapt==1.12.1
